move to .Extensions namespace
