+++ src/Sidekick.Apis.Poe/Clients/PoeTradeClient.cs
@@ -2,6 +2,7 @@ using System.Net;
 using Microsoft.Extensions.Logging;
 using Sidekick.Apis.Poe.Clients.Models;
 using Sidekick.Common;
+using Sidekick.Common.Extensions;
 using Sidekick.Common.Game;
 using Sidekick.Common.Game.Languages;
